i2c: i2c-msm-geni: add i2c error check for fifo mode
In FIFO mode for RTL-based SE, whenever an I2C transfer
times out, the I2C GENI driver will check the status of
the I/O lines. If the I/O lines are not in a good state,
it marks a pending cancel and returns without mentioning
any error.

If we do not return with a proper error code, it can lead
to undetected issues and causing potential data corruption
or loss and making debugging difficult.

To address this, an error check is added at the end of the
transfer.

The geni_i2c_execute_xfer function contains multiple
condition checks, resulting in high cyclomatic complexity.
To avoid these errors and improve readability, the code
in the geni_i2c_execute_xfer function has been modularized
into separate function.

Change-Id: I42a9e570911ae37e3e9eb934be3d47d6cf10f6fc
Signed-off-by: Anil Veshala Veshala <quic_aveshala@quicinc.com>