sched: walt: sanity checks around wts->grp ptr
wts->grp is used to indicate if a task belongs to related thread group.
There are 20 groups with group 1 DEFAULT_CGROUP_COLOC_ID,  tracking tasks
in top-app cgroup.

Group 0 is technically unused but the element number 0 is used to
denote that the task doesn't belong to any related thread group.

Enable a change to ensure that only group 1 is used if wts->grp is
set. Also disable the ability to use other groups i.e. disable
the sysfs node that allows  to use group other than 1.

This will help us track any corruptions of the grp ptr which could
lead to a task incorrectly adding/removing its busy time to grptime.

This change can be reverted once the accounting problem is addressed
and a need for multiple related thread group is needed.

Change-Id: I2cd49ef98f191815fd6c11b76a72651077d67949
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Srinivasarao Pathipati <quic_c_spathi@quicinc.com>